Moving unblocking usage of local AGW IPs from configure to install
